pyproject.toml
[project]
name = "fastapi-url-shortener"
version = "0.1.0"
description = "FastAPI URL Shortener"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "faststream[rabbit]>=0.5.48",
    "jinja2>=3.1.6",
    "pika>=1.3.2",
    "pydantic-settings[yaml]>=2.10.1",
    "python-multipart>=0.0.20",
    "rabbitmq>=0.2.0",
    "redis[hiredis]>=5.2.1",
    "requests>=2.32.5",
    "setuptools>=80.9.0",
    "typer>=0.15.2",
]

[dependency-groups]
dev = [
    "coverage>=7.10.1",
    "httpx>=0.28.1",
    "mypy>=1.17.0",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.11.5,<=0.13.5",
    "stripe>=12.5.1",
]

[tool.uv]
required-version = "~=0.7"


[tool.pytest.ini_options]
minversion = "8.3"
addopts = "-ra --strict-markers"
log_cli = false
log_cli_level = "INFO"
xfail_strict = true

markers = [
    "apitest: test any http api call",
    "templatetest: test http HTML template page"
]



# mypy settings

[tool.mypy]
strict = true

plugins = ["pydantic.mypy"]



#roof settings

[tool.ruff]
required-version = ">=0.11.5,<=0.13.5"

src = [
    "url-shortener",
]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.13
target-version = "py313"

[tool.ruff.lint]
select = [
    "FAST", # FastAPI (FAST)
    "YTT", # flake8-2020 (YTT)
    "ANN", # flake8-annotations (ANN)
    "BLE", # flake8-blind-except (BLE)
    "FBT", # flake8-boolean-trap (FBT)
    "C4", # flake8-comprehensions (C4)
    "DTZ", # flake8-datetimez (DTZ)
    "T10", # flake8-debugger (T10)
    "EM", # flake8-errmsg (EM)
    "ISC", # flake8-implicit-str-concat (ISC)
    "LOG", # flake8-logging (LOG)
    "G", # flake8-logging-format (G)
    "PIE", # flake8-pie (PIE)
    "PYI", # flake8-pyi (PYI)
    "Q", # flake8-quotes (Q)
    "RSE", # flake8-raise (RSE)
    "RET", # flake8-return (RET)
    "SLF", # flake8-self (SLF)
    "SIM", # flake8-simplify (SIM)
    "SLOT", # flake8-slots (SLOT)
    "TC", # flake8-type-checking (TC)
    "PTH", # flake8-use-pathlib (PTH)
    "FLY", # flynt (FLY)
    "I", # isort (I)
    "PGH", # pygrep-hooks (PGH)
    "PLC", # Pylint Convention (PLC)
    "PLE", # Pylint Error (PLE)
    "PLR", # Pylint Refactor (PLR)
    "PLW", # Pylint Warning (PLW)
    "FURB", # refurb (FURB)
]
ignore = [
    "S101", # S101	assert	Use of assert detected
    "RUF001", # ambiguous-unicode-character-string (RUF001)
    "RUF002", # ambiguous-unicode-character-docstring (RUF002)
    "RUF003", # ambiguous-unicode-character-comment (RUF003)
    "TC006", # runtime-cast-value (TC006)
    "TRY", # tryceratops (TRY)
    "DOC", # pydoclint (DOC)
    "D", # pydocstyle (D)
    "PT", # flake8-pytest-style (PT)
    "ASYNC", # flake8-async (ASYNC)
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []
